+++ src/ChapterNavigation.php
@@ -20,11 +20,12 @@ class ChapterNavigation
 			'order' => 'ASC'
 		]);
 		array_unshift($chapterPosts, $parentPost);
+		$currentPostId = $post->ID;
 		$chapterNavigationItems = [];
 		foreach ($chapterPosts as $chapterPost) {
 			$chapterNavigationItems[] = (object) [
 				'title' => $chapterPost->post_title,
-				'url' => $chapterPost->ID == $post->ID ? null : get_permalink($chapterPost)
+				'url' => apply_filters('long_read_plugin_chapter_url', $chapterPost->ID == $currentPostId ? null : get_permalink($chapterPost), $chapterPost->ID)
 			];
 		}
 		return $chapterNavigationItems;